Move alias of ProtocolKeys to a wrapped module
