+++ mithril-common/src/crypto_helper/types/mod.rs
@@ -1,8 +1,8 @@
 mod alias;
 #[macro_use]
 mod protocol_key;
-mod single_signature;
+mod wrappers;
 
 pub use alias::*;
 pub use protocol_key::*;
-pub use single_signature::*;
+pub use wrappers::*;
