--- mithril-common/src/crypto_helper/types/single_signature.rs
@@ -1,6 +1,0 @@
-use crate::crypto_helper::ProtocolKey;
-use mithril_stm::stm::StmSig;
-
-/// Wrapper of [MithrilStm:StmSig](type@StmSig)
-/// to add serialization utilities.
-pub type ProtocolSingleSignature = ProtocolKey<StmSig>;
